Explicitly add Map types to avoid circular dependencies
