+++ packages/react-ui/src/index.ts
@@ -14,3 +14,8 @@ export * from './Icons'
 export * from './PopUps'
 export * from './Tables'
 export * from './Map/useMapBaseLayer'
+export type { MapProps } from './Map/Map.types'
+export type {
+  MapDepthDisplayProps,
+  DepthRequestFn,
+} from './Map/MapDepthDisplay'
